+++ hami-core/src/main/java/top/wang3/hami/core/mapper/ArticleMapper.java
@@ -45,6 +45,6 @@ public interface ArticleMapper extends BaseMapper<Article> {
 
     List<ArticleCount> selectCateArticleCount();
 
-    Long selectTotalArticleCount();
+    Integer selectTotalArticleCount();
 
 }
