+++ hami-core/src/main/java/top/wang3/hami/core/service/article/repository/ArticleRepository.java
@@ -37,7 +37,7 @@ public interface ArticleRepository extends IService<Article> {
 
     Integer getArticleAuthor(Integer articleId);
 
-    Map<String, Long> getArticleCount();
+    Map<String, Integer> getArticleCount();
 
     boolean saveArticle(Article article);
