+++ hami-core/src/main/java/top/wang3/hami/core/service/interact/LikeService.java
@@ -13,7 +13,7 @@ public interface LikeService {
 
     boolean cancelLike(Integer itemId, LikeType likeType);
 
-    Long getUserLikeCount(Integer userId, LikeType likeType);
+    Integer getUserLikeCount(Integer userId, LikeType likeType);
 
     List<Integer> listUserLikeArticles(Page<LikeItem> page, Integer userId);
