+++ hami-core/src/main/java/top/wang3/hami/core/service/interact/repository/FollowRepository.java
@@ -11,17 +11,19 @@ public interface FollowRepository extends IService<UserFollow> {
 
     /**
      * 获取用户关注数
+     *
      * @param userId 用户ID
      * @return 用户关注数
      */
-    Long getUserFollowingCount(Integer userId);
+    Integer getUserFollowingCount(Integer userId);
 
     /**
      * 获取用户粉丝数
+     *
      * @param userId 用户ID
      * @return 用户粉丝数
      */
-    Long getUserFollowerCount(Integer userId);
+    Integer getUserFollowerCount(Integer userId);
 
     boolean hasFollowed(Integer userId, Integer followingId);
